refactor: centralize prompt template loading logic (T207)
Consolidate duplicate template loading logic into a single implementation:
- Move prompt loading to config/template.go with a clear, exported API
- Remove duplicate logic from llm/prompt.go
- Mark old LoadTemplate function as deprecated
- Make LLM service use templates from config only, no more direct file loading
- Add dedicated template tests and update existing tests
- Skip integration tests that need to be fixed separately

Now the config package owns all template loading with proper validation.
