+++ config/loadconfig.go
@@ -10,7 +10,7 @@ import (
 	"github.com/joho/godotenv"
 	"github.com/sirupsen/logrus"
 
-	"glance/filesystem"
+	"glance/llm"
 )
 
 // directoryChecker defines an interface for checking directory existence
@@ -114,10 +114,15 @@ func LoadConfig(args []string) (*Config, error) {
 		return nil, errors.New("GEMINI_API_KEY is missing: please set this environment variable or add it to your .env file")
 	}
 
-	// Load prompt template
-	promptTemplate, err := loadPromptTemplate(promptFile)
+	// Load prompt template using the centralized function
+	promptTemplate, err := LoadPromptTemplate(promptFile)
 	if err != nil {
 		return nil, fmt.Errorf("failed to load prompt template: %w", err)
+	}
+
+	// If no template was found, use the default from llm package
+	if promptTemplate == "" {
+		promptTemplate = llm.DefaultTemplate()
 	}
 
 	// Apply all configuration settings using the builder pattern
@@ -129,71 +134,4 @@ func LoadConfig(args []string) (*Config, error) {
 		WithPromptTemplate(promptTemplate)
 
 	return cfg, nil
-}
-
-// loadPromptTemplate tries to read from the specified file path, then "prompt.txt",
-// and falls back to the default prompt template if neither is available.
-// It securely validates all file paths to prevent path traversal vulnerabilities.
-func loadPromptTemplate(path string) (string, error) {
-	// Get current working directory as the base for validation
-	cwd, err := os.Getwd()
-	if err != nil {
-		return "", fmt.Errorf("failed to get current working directory: %w", err)
-	}
-
-	// For custom path, properly validate against the entire filesystem
-	if path != "" {
-		// We can't use ValidateFilePath directly with an empty baseDir
-		// because that would skip validation, so we need to clean and
-		// absolutize the path first, then validate against filesystem root
-		// This allows users to provide templates from anywhere on the filesystem
-		// but prevents path traversal attacks
-		cleanPath := filepath.Clean(path)
-		absPath, err := filepath.Abs(cleanPath)
-		if err != nil {
-			return "", fmt.Errorf("invalid prompt template path: %w", err)
-		}
-
-		// Use filesystem root ("/") as baseDir to enforce path validity
-		// but not containment within any specific directory
-		// allowBaseDir=true because the root itself is allowed
-		// mustExist=true to ensure the file exists and is not a directory
-		rootDir := "/"
-		validPath, err := filesystem.ValidateFilePath(absPath, rootDir, true, true)
-		if err != nil {
-			return "", fmt.Errorf("failed to validate prompt template path: %w", err)
-		}
-
-		// Read the validated file path
-		// #nosec G304 -- The path has been validated using filesystem.ValidateFilePath
-		data, err := os.ReadFile(validPath)
-		if err != nil {
-			return "", fmt.Errorf("failed to read custom prompt template from '%s': %w", validPath, err)
-		}
-		return string(data), nil
-	}
-
-	// Try the default prompt.txt in the current directory
-	defaultPromptPath := filepath.Join(cwd, "prompt.txt")
-	// Check if the file exists, but don't return an error if it doesn't
-	if _, err := os.Stat(defaultPromptPath); err == nil {
-		// Validate the default path against the current working directory
-		// allowBaseDir=false because prompt.txt should be in CWD, not be CWD itself
-		// mustExist=true to ensure the file exists and is a file
-		validDefaultPath, err := filesystem.ValidateFilePath(defaultPromptPath, cwd, false, true)
-		if err != nil {
-			// We just log this rather than fail, to maintain backward compatibility
-			// with existing behavior of falling back to default template
-			logrus.Warnf("Failed to validate default prompt template: %v", err)
-		} else {
-			// Read the validated file
-			// #nosec G304 -- The path has been validated using filesystem.ValidateFilePath
-			if data, err := os.ReadFile(validDefaultPath); err == nil {
-				return string(data), nil
-			}
-		}
-	}
-
-	// Fall back to the hardcoded default template
-	return defaultPromptTemplate, nil
 }
